(VANASGON-125) Stop services in the correct place on AIX upgrades
Before the work for VANAGON-120, services from puppet-agent were
restarted in the postun step. This caused the following sequence of
events:

1. pre - collect service status
2. post - [noop]
3. preun - stop the service
4. postun - restart the service if needed

Now that service restart has been moved to the post, stopping in the
preun is broken for two reasons:

1. We do not actually stop the service before trying to restart it
2. The service is stopped after we have nominally restarted it

This makes the stop in preun conditional, so that it only runs on true
uninstalls and not on upgrades. It also adds a new stop in pre, after
the project-specific preinstall script, to stop the service.