(VANAGON-125) Move AIX postinstall scripts to 'triggerpostun'
In VANAGON-120, we moved postinstall upgrade scripts from the postun
step to the post step, so that they would always come from the new
version of the package, instead of the old. Unfortunately, vanagon
historically has had an unconditional service stop in the preun for
AIX. This meant that when upgrading from older packages, the services
we restart in post would be stopped again in preun.

We need to ensure we can still run the version of the script that
comes with the new package, but do so after the old package's
aggressive preun has run. To do that, the new package registers a
triggerpostun handler for its own name. This handler will be run from
the new package near the end of the transaction - after the old
package's preun and postun scripts.